requirements.txt
# Core Deep Learning Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# Video Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ffmpeg-python>=0.2.0

# Face Detection
facenet-pytorch>=2.5.3
mtcnn>=0.1.1

# Image Processing
Pillow>=10.0.0
albumentations>=1.3.1

# Pre-trained Models
timm>=0.9.2
efficientnet-pytorch>=0.7.1

# Data Handling
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
tensorboard>=2.13.0
tqdm>=4.65.0

# Utilities
PyYAML>=6.0
requests>=2.31.0

# GPU Monitoring
gpustat>=1.1